requirements.txt
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.1
beautifulsoup4>=4.9.3
requests>=2.25.1
python-dotenv>=0.19.0
pandas>=1.3.0
selenium>=4.0.0
webdriver_manager>=3.5.0
linkedin-api>=2.0.0
validators>=0.18.2
logging>=0.5.1.2